+++ examples/getting_started.py
@@ -43,7 +43,7 @@ class BasicApp(cmd2.Cmd):
         shortcuts = cmd2.DEFAULT_SHORTCUTS
         shortcuts.update({'&': 'intro'})
         super().__init__(
-            include_bottom_toolbar=True,
+            bottom_toolbar=True,
             include_ipy=True,
             multiline_commands=['echo'],
             persistent_history_file='cmd2_history.dat',